firmware: add new file: compilermacros.h
GCC rightly warns us of unused arguments. While it's possible to work around
intentionally unused arguments with constructions such as "(void)argname" in the
function body, these only work a) by accident or b) because compilers such as
GCC explicitly recognise this idiom.

Most compilers have some implementation-dependent way of indicating an argument
is intentionally unused. Wrapping this in a macro allows us to a) remain
mostly compiler agnostic and b) explicitly indicate to future humans reading the
code that we intentionally ignore the argument.