firmware:mpu9250: add gyro self-test to init sequence
Re-jig mpu9250_init() to use a table of registers and values to make it a little
easier to edit. Also call (and check return value of) mpu9250_self_test_gyro()
to perform a self-test on the gyroscope.

It's non-obvious where to find the application note referenced in the datasheet
on self-testing. Instead I've cribbed from other MPU9250 drivers and implemented
the same self-test algorithm.

Essentially the driver now takes a mean measurement before and after self-test
activation and subtracts them to get the "pure" self-test effect. Then the
factory-set expected values are expanded out from the appropriate registers. If
the self-test values differ from the expect values by more than 5% the self-test
fails. I've chosen the 5% threshold somewhat arbitrarily.

A little bit of work is done to minimise the number of floating point operations
which need to be performed. Most of the work is kept in the integer domain.