firmware: disable OTG VBUS sense support
There are a plethora of ways to wire the STM32F4 up for USB OTG support. In our
case we are wanting peripheral-only mode and, importantly, without use of VBUS
sense. (This frees PA9 for other uses.)

Without BOARD_OTG_NOVBUSSENS defined, ChibiOS configures the MCU to use VBUS
sense support and since the PA9 line is probably unconnected, it will never
detect a USB cable as being plugged in.