firmware:mpu9250: add minimal AK8963 information
On the MPU9250, the magnetometer is a separate AK8963 within the same package.
From peering at example code (and datasheets) it appears that this is controlled
at arm's length via the MPU9250's I2C slave support.

Note: this means out current mpu9250_burst_read() implementation is wrong.

Add a list of AK8963 registers and the I2C address of the AK8963 packaged within
the MPU9250.